requirements.txt
pycryptodomex == 3.20.0
pwinput == 1.0.3
PyQT5 == 5.15.10
